requirements.txt
# Google ADK and Gemini
google-adk>=0.1.0
google-genai>=0.2.0

# PDF Processing
pypdf>=3.17.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0

# Embeddings and Vector Store
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.24.0

# Internet Search - Now using Gemini's integrated Google Search
# No external search API dependencies needed

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.5.0
